app.py
## RAG Q&A Conversation With PDF Including Chat History

import streamlit as st
from langchain_classic.chains import create_history_aware_retriever, create_retrieval_chain
from langchain_classic.chains.combine_documents import create_stuff_documents_chain
from langchain_chroma import Chroma
from langchain_community.chat_message_histories import ChatMessageHistory
from langchain_core.chat_history import BaseChatMessageHistory
from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
from langchain_groq import ChatGroq
from langchain_core.runnables.history import RunnableWithMessageHistory
from langchain_huggingface import HuggingFaceEmbeddings
from langchain_community.document_loaders import PyPDFLoader
from langchain_text_splitters  import RecursiveCharacterTextSplitter
import os
from dotenv import load_dotenv
load_dotenv()

os.environ["HF_TOKEN"] = os.getenv("HF_TOKEN")
embeddings=HuggingFaceEmbeddings(model_name="sentence-transformers/all-MiniLM-L6-v2")

# Set up Streamlit app
st.title("RAG Q&A Assistant with Chat History - PDF Uploads")
st.write("Upload Pdf's and chat with their content")

# input groq API key
api_key = st.text_input("Enter your Groq API Key:", type="password")

# Check if the API key is provided
if api_key:
    # Initialize the ChatGroq model
    llm = ChatGroq(
        groq_api_key=api_key,
        model_name="llama-3.3-70b-versatile"
    )

    # chat interface
    session_id = st.text_input("Session ID", value="default_session")

    if 'store' not in st.session_state:
        st.session_state.store = {}

    uploaded_files = st.file_uploader("Upload PDF files", accept_multiple_files=True, type=["pdf"])

    # Process the uploaded PDF files
    if uploaded_files:
        documents=[]
        for uploaded_file in uploaded_files:
            temppdf=f"./temp.pdf"
            with open(temppdf, "wb") as f:
                f.write(uploaded_file.getvalue())
                f_name = uploaded_file.name
            
            loader = PyPDFLoader(temppdf)
            docs = loader.load()
            documents.extend(docs)

        # Split and create embeddings for the documents
        text_splitter = RecursiveCharacterTextSplitter(chunk_size=5000, chunk_overlap=500)
        splits = text_splitter.split_documents(documents)
        vectorstore = Chroma.from_documents(splits, embeddings, collection_name=session_id)
        retriever = vectorstore.as_retriever()

        ## Create a prompt template for question contextualization
        contextualize_q_system_prompt = (
            """
            Given a chat history and the latest user question, which might reference context in the \
            chat history, formulate a standalone question which can be understood without the chat history. Do NOT answer the \
            question, just reformulate it if needed and otherwise return the question as is. \
            """
        )

        context_q_prompt = ChatPromptTemplate.from_messages(
            [
                ("system", contextualize_q_system_prompt),
                MessagesPlaceholder("chat_history"),
                ("human", "{input}")
            ]
        )

        history_aware_retriever=create_history_aware_retriever(llm, retriever, context_q_prompt)

        ## Answer question 
        system_prompt = (
            """
            You are a helpful assistant for answering questions about the content of the provided documents. \
            Use only the provided context to answer the question. If you don't know the answer, say you don't know. \
            Use three sentences maximumand keep the answer concise. \
            \n\n
            {context}
            """
        )

        qa_prompt = ChatPromptTemplate.from_messages(
            [
                ("system", system_prompt),
                MessagesPlaceholder("chat_history"),
                ("human", "{input}")
            ]
        )

        question_answer_chain = create_stuff_documents_chain(llm, qa_prompt)
        rag_chain = create_retrieval_chain(history_aware_retriever, question_answer_chain)

        def get_session_history(session_id: str) -> BaseChatMessageHistory:
            if session_id not in st.session_state.store:
                st.session_state.store[session_id] = ChatMessageHistory()
            return st.session_state.store[session_id]
        
        conversation_rag_chain = RunnableWithMessageHistory(
            rag_chain, 
            get_session_history,
            input_messages_key="input",
            history_messages_key="chat_history",
            output_messages_key="answer"
        )

        user_input = st.text_input("Ask a question about the documents:")
        if user_input:
            session_history = get_session_history(session_id)
            response = conversation_rag_chain.invoke(
                {"input": user_input}, 
                config={"configurable": {"session_id": session_id}}
            )
            st.write(st.session_state.store)
            st.write("Answer:", response['answer'])
            st.write("Chat History:", session_history.messages)
else:
    st.warning("Please enter your Groq API Key to use the application.")